Add the term networkRequirements
This adds the term networkRequirements as described in #191

I've checked the various endpoints of the crosswalk and can not find
a case in which this term is expressed, so the tables is blank -
only codemeta will have this term.

It is needed for the proper description of infrastructure products
such as grid and cloud middleware.

The term takes a `Text` value:

  - IPv4 (for IPv4 support only)
  - IPv6 (for IPv6 support only)
  - dualStack (for dual stack support)